# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domain.Instance:
    properties:
      container_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      ports:
        type: string
      status:
        $ref: '#/definitions/domain.InstanceStatus'
      updated_at:
        type: string
      version:
        type: integer
      vpc_id:
        type: string
    type: object
  domain.InstanceStats:
    properties:
      cpu_percentage:
        type: number
      memory_limit_bytes:
        type: number
      memory_percentage:
        type: number
      memory_usage_bytes:
        type: number
    type: object
  domain.InstanceStatus:
    enum:
    - STARTING
    - RUNNING
    - STOPPED
    - ERROR
    - DELETED
    type: string
    x-enum-varnames:
    - StatusStarting
    - StatusRunning
    - StatusStopped
    - StatusError
    - StatusDeleted
...
  httphandlers.LaunchRequest:
    properties:
      image:
        type: string
      name:
        type: string
      ports:
        type: string
      volumes:
        items:
          $ref: '#/definitions/httphandlers.VolumeAttachmentRequest'
        type: array
      vpc_id:
        type: string
    required:
    - image
    - name
    type: object
  httphandlers.VolumeAttachmentRequest:
    properties:
      mount_path:
        type: string
      volume_id:
        type: string
    type: object
  httputil.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  httputil.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/httputil.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Mini AWS Compute API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mini AWS API
  version: "1.0"
paths:
...
  /instances:
    get:
      description: Gets a list of all compute instances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Instance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: List all instances
      tags:
      - instances
    post:
      consumes:
      - application/json
      description: Creates and starts a new compute instance with optional volumes
        and VPC
      parameters:
      - description: Launch request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.LaunchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Instance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Launch a new instance
      tags:
      - instances
  /instances/{id}:
    delete:
      description: Deletes a compute instance and its associated resources
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Terminate an instance
      tags:
      - instances
    get:
      description: Gets detailed information about a specific compute instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Instance'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get instance details
      tags:
      - instances
  /instances/{id}/logs:
    get:
      description: Gets the console output logs for a compute instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Logs content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get instance logs
      tags:
      - instances
  /instances/{id}/stats:
    get:
      description: Gets real-time CPU and Memory usage for a compute instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.InstanceStats'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get instance stats
      tags:
      - instances
  /instances/{id}/stop:
    post:
      description: Initiates a graceful shutdown of a compute instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Stop an instance
      tags:
      - instances
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"